Introduce Address type
Address now can be either Contract or Implicit
I removed pk field because it's not used anywhere and
is impossible to get for contracts (unification)

Now we are able to restrict where we can use only Contracts or
only Implicits, or both. parsePkh function can be further improved
to validate contracts and make sure we have valid addresses throughout
the app.

This change allows to smoothly introduce other types of contract
accounts

As for now we don't support Multisig hierarchies and this change
checks that we don't allow to pass in contract accounts as signers.