Make SignButton manage TezosToolkit build process
This way one a user hits the sign button we have a TezosToolkit
ready to sign and there is no need to repeat the same process down
the line in operations.ts.

Also, to prepare the codebase for multisig batch estimations I pass
both the sender and the signer.

Alongside that useApproveOrExecute is not needed anymore because it's
functionality is covered by the SignButton.

The Address & Account types are used instead of RawPkh where possible
so that we don't need to keep pulling it from the state.