Introduce new way to handle operations
Now we have a global modal placement that can be used from
anywhere in the app.

Going back is done using the same `openWith` callback.
If page A opens page B and page B should have the go back button
then page A should let page B how to go back to page A (see Tez
send form for the example)

Modals are now centered and are much closer to the design

AddressAutocomplete can be used as a selectbox until we
implement one. It's done that way so that we cannot have an
missing signer on the signer page.

Sign page has a signer field for multisig operations and if
the signer changes it will reestimate the operation and let the user
know if that's impossible to sign the operation with the new signer

Sign button text is unified for the new sign page, but remains the same
for the former one.