Migrate app to ES modules
That was the original option we used and the only
reason it was changes is because cucumber requires
CommonJS modules instead to be able to define and
run the tests.

Now we have a separate tsconfig for cucumber only
so that it can stick to CJS modules while the app
is using ES modules.

This will allow us also to migrate to Vite easily
because CJS are deprecated there